requirements.txt
Flask==2.0.3
google-cloud-storage==1.43.0
PyCryptodome==3.10.1
speechrecognition==3.8.1
